feat(vue-demo-store): Custom registration page
