+++ packages/types/shopware-6-client/models/checkout/customer/CustomerGroup.d.ts
@@ -8,4 +8,18 @@ export type CustomerGroup = {
   name: string;
   displayGross: boolean;
   customFields: CustomFields;
+  createdAt: string;
+  updatedAt: null | string;
+  displayGross: boolean;
+  registrationActive: boolean;
+  registrationTitle: string;
+  registrationIntroduction: string;
+  registrationOnlyCompanyRegistration: boolean;
+  registrationSeoMetaDescription: null | string;
+  translated: {
+    [key in keyof Omit<
+      CustomerGroup,
+      "translated"
+    > as CustomerGroup[key] extends string ? key : never]: string;
+  };
 };
